Get Cited By PMIDs from PubMed

List PubMed IDs of publications citing those specified by web_history or id, potentially split by_id. citedby_pmid() can handle cases where the number of input IDs is >200 automatically (unlike rentrez::entrez_link).

Usage

citedby_pmid(id = NULL, web_history = NULL, by_id = FALSE, config = NULL, ...)

Arguments

id

PubMed IDs for which to retrieve "cited by", as a character vector.

web_history

An Entrez History object to be used instead of id, as generated by the modified rentrez::entrez_post() fork at https://github.com/allenbaron/rentrez (see NOTE); should be used when "cited by" information is desired for >200 IDs.

by_id

Whether to split "cited by" publications by the ID they cite, as a boolean (default: TRUE); FALSE returns a unified list of unique cited by publications without reference to the ID they cite.

config

Configuration options passed to httr::GET().

...

Additional terms to add to the request, see NCBI documentation on ESearch for a complete list.

NOTE

The rentrez fork at https://github.com/allenbaron/rentrez implements pull request #174, which is not yet implemented in the CRAN rentrez and is required for rentrez::entrez_post() to function fully.