Get Cited By List from PubMed

List PubMed publications citing those specified by id, optionally split by_id.

Usage

citedby_pubmed(
  id = NULL,
  web_history = NULL,
  by_id = FALSE,
  no_result = "warning",
  config = NULL,
  version = "2.0",
  retmode = "xml",
  ...
)

Arguments

id

PubMed IDs for which to retrieve "cited by", as a character vector.

web_history

An Entrez History object to be used instead of id, as generated by the modified rentrez::entrez_post() fork at https://github.com/allenbaron/rentrez (see NOTE); should be used when "cited by" information is desired for >200 IDs.

by_id

Whether to split "cited by" publications by the ID they cite, as a boolean (default: TRUE); FALSE returns a unified list of unique cited by publications without reference to the ID they cite.

no_result

The type of condition that should be signaled when no PubMed results exist in a response; one of "error", "warning" (default), "message" or "none".

config

Configuration options passed to httr::GET().

version

Argument included here for flexibility, but best left with the defaults as set; see rentrez::entrez_summary() for details.

retmode

"xml" (default) or "json"; "xml" is preferred because of a higher response limit. This default is opposite the rentrez::entrez_summary() default.

...

Additional terms to add to the request, see NCBI documentation on ESearch for a complete list.

See also