Parse Mapping

Parses mappings generated by pyobo_map() or otherwise produced by the python module GILDA.

Usage

parse_mapping(py_gilda_res_list, ...)

Arguments

py_gilda_res_list

A named list of term mapping lists (see parse_term_mapping()), likely generated by pyobo_map().

...

Arguments passed on to extract_ScoredMatch, parse_term_mapping

prefix

Optional prefix to add to namespace local unique identifiers (LUI; e.g. 4, the LUI for "disease" in DO), as a string; preferably to create a complete namespace ID (e.g. "DOID:4").

prefix_sep

Optional separator placed between prefix and namespace LUIs, as a string. Ignored if prefix = NULL.

best_only

Whether to return the best scoring result(s) only, as a boolean. FALSE will return all results. NOTE: TRUE will return more than one result in the case of ties.

warn_best_gt1

Whether to warn that best mapping ties exist for a term.

Value

List of data frames with the mapping result(s) for each term.