Instantiate a DO Repository

Instantiate a DOrepo object from the pyDOID package (powered by reticulate).

Usage

DOrepo(path)

Arguments

path

The local path to the HumanDiseaseOntology git repository, as a string.

Classes & Methods

The DOrepo class inherits the methods & subclasses of the GitPython.Repo class, along with the following:

DOrepo$tag_iterate(function, start = [str], end = [str]): Iterate through DO releases (tags) from start to end executing a function at each. start and end are optional; if unspecified the first or last release/tag of the repo will be assumed.

DOrepo$capture_head(): Capture the current head of the repo; useful for restoring state after git checkout(s). USE: If assigned to variable x, restore state with x$checkout().

DOrepo$doid and DOrepo$doid_merged: Access the doid.owl and doid-merged.owl files of the repository. These inherit from the pyDOID.owl.xml class, see owl_xml() for methods. NOTE that, unlike owl_xml(), doid.owl and doid-merged.owl are not loaded into memory upon instantiation of DOrepo, but are loaded automatically when the first $query() is executed or manually via $load().

See also